Fix IntLongMath overflow warning in TProtocol
Cast to long before addition to prevent potential integer overflow.

+++ lib/java/src/main/java/org/apache/thrift/protocol/TProtocol.java
@@ -78,7 +78,7 @@ public abstract class TProtocol implements TWriteProtocol, TReadProtocol {
   }
 
   protected void checkReadBytesAvailable(TMap map) throws TException {
-    long elemSize = getMinSerializedSize(map.keyType) + getMinSerializedSize(map.valueType);
+    long elemSize = (long) getMinSerializedSize(map.keyType) + getMinSerializedSize(map.valueType);
     trans_.checkReadBytesAvailable(map.size * elemSize);
   }
